+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - octave & toOctave number extensions
 
+- root & pitches properties added to IPitchContainer
+
+- bass property added to Chord
+
 
 ## 2.6.0 (2023-11-10)